[backup_influxdb] Execute backup within shell
The direct execution of backup commands leads to 401/unauthorizes for the metadata.
If the backup is run within a bash, it works well.

Regarding the deletion, only files within the InfluxDB backup directories should be deleted.
Otherise, the whole backup directory may get lost if there are no current backups.